% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{areColors}
\alias{areColors}
\title{Check if color is valid}
\usage{
areColors(x)
}
\arguments{
\item{x}{A color as a string}
}
\value{
A logical TRUE or FALSE depending it the color string is valid.
}
\description{
Check if color is valid
}
