% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrama.R
\name{ggrama}
\alias{ggrama}
\title{Ramachandran diagram}
\usage{
ggrama(
  pdb,
  type = c("general", "glycine", "pre.pro", "proline"),
  values = switch(type, general = c(0, 0.001, 0.02, 0.1, 1), glycine = c(0, 0.002,
    0.02, 0.1, 1), proline = c(0, 0.002, 0.02, 0.1, 1), pre.pro = c(0, 0.002, 0.02, 0.1,
    1)),
  colors = switch(type, general = c("#FFFFFF", "#B3E8FF", "#82daff", "#19bbff"),
    glycine = c("#FFFFFF", "#FFE8C5", "#ffddab", "#FFCC7F"), proline = c("#FFFFFF",
    "#D0FFC5", "#a2fa8e", "#49de26"), pre.pro = c("#FFFFFF", "#B3E8FF", "#82daff",
    "#19bbff")),
  title = switch(type, general = "General", glycine = "Glycine (Symmetric)", proline =
    "Proline", pre.pro = "Pre-Proline"),
  shape = 21,
  size = 2,
  fill = "#FFFFFF",
  contour = FALSE,
  contour.color = NA,
  smooth = FALSE,
  smoothing.param = c("2", "3", "4"),
  ...
)
}
\arguments{
\item{pdb}{input pdb file location.}

\item{type}{One of c("general", "glycine", "pre.pro", "proline"). String representing the type of Ramachandran diagram.}

\item{values}{Numeric vector indicating the values of \code{\link{ggplot2}{scale_fill_gradientn}}.}

\item{colors}{A character vector of valid color names indicating the colors for \code{\link{ggplot2}{scale_fill_gradientn}}.}

\item{title}{A string indicating the title of the plot.}

\item{shape}{Shape for \code{\link{ggplot2}{geom_point}}.}

\item{size}{Size for \code{\link{ggplot2}{geom_point}}.}

\item{fill}{fill for \code{\link{ggplot2}{geom_point}} if supported.}

\item{contour}{logical, should  \code{\link{ggplot2}{geom_contour_filled}} be used instead of \code{\link{ggplot2}{geom_tile}}}

\item{contour.color}{numeric, the color of the contour if contour = TRUE.}

\item{smooth}{logical, should the background grid be interpolated.}

\item{smoothing.param}{character - "2", "3" or "4", the coefficient of interpolation. Higher values give smoother contours at the expense of computation time.}

\item{...}{currently not used}
}
\value{
A ggplot2 plot object
}
\description{
Plots a Ramachandran diagram from a pdb file using ggplot2 graphics
}
\examples{

#download a PDB file
download.file("https://files.rcsb.org/download/2ADQ.pdb1.gz",
              dest = "2ADQ.pdb1.gz")

# plot the general Ramachandran diagram
ggrama("2ADQ.pdb1.gz")

# plot the proline Ramachandran diagram
ggrama("2ADQ.pdb1.gz", type = "proline")

}
