% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrama_all.R
\name{ggrama_all}
\alias{ggrama_all}
\title{Ramachandran diagrams}
\usage{
ggrama_all(pdb, ...)
}
\arguments{
\item{pdb}{input pdb file location.}

\item{...}{arguments passed to  \code{\link{ggrama}{ggrama}}}
}
\value{
a grid of four ggplot2 figures combined using  \code{\link{cowplot}{plot_grid}}
}
\description{
Plots four types of Ramachandran diagrams in a grid from a pdb file using ggplot2 graphics
}
\examples{

#download a PDB file
download.file("https://files.rcsb.org/download/2ADQ.pdb1.gz",
              dest = "2ADQ.pdb1.gz")

#plot
ggrama_all("2ADQ.pdb1.gz")

}
