% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{resize_mat_linear}
\alias{resize_mat_linear}
\title{1D linear matrix resize}
\usage{
resize_mat_linear(img, new_width, new_height)
}
\arguments{
\item{img}{input matrix}

\item{new_width}{output matrix width}

\item{new_height}{output matrix height}
}
\value{
resized matrix of dimensions new_height * new_width
}
\description{
1D linear matrix resize
}
